+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 [Unreleased]: https://github.com/atomist/rug/compare/0.26.1...HEAD
 
+### Changed
+
+-   **BREAKING** TypeScript `Match` now uses property access, not function
+
 ## [0.26.1] - 2017-04-24
 
 [0.26.1]: https://github.com/atomist/rug/compare/0.26.0...0.26.1